security: comprehensive security audit fixes (31 issues) + bump to 0.2.15
High:
- Rust: handle_import_full path traversal (key_path canonicalize + allowlist)
- Rust: handle_import_full deadlock (drop Mutex guard before re-lock)
- Rust: shell_escape breaking firewalld/telegram templates (shell var indirection)
- Android: private key file 0600 perms + cleanup on disconnect/remove
- CI: pin all GitHub Actions to commit SHA
- CI: cloud-sync.php exchange endpoint redirect_uri whitelist

Medium:
- Rust: import_full bind magic as GCM AAD
- Android: Rust logcat default Warn in release + nativeSetLogLevel
- Android: MainActivity Intent action validation (reject third-party VPN trigger)
- Android: APK SHA-256 in release body + extractSha256FromBody
- Android: remove unused REQUEST_INSTALL_PACKAGES permission
- CI: ip.php/ip2.php prefer REMOTE_ADDR over XFF
- CI: release.yml top-level permissions block
- CI: npm ci instead of npm install
- CI: validate TERMFAST_ secrets in CI
- CI: fs:scope deny list expanded (.netrc/.kube/.docker/.npmrc/.git-credentials)
- CI: shell:allow-open scoped to https://*/*

Low:
- Android: BuildConfig.DEBUG gate on Log.i with sensitive data
- Android: remove unused RECEIVE_BOOT_COMPLETED permission
- Android: cleanup credentials + key dir on server removal
- Android: master password min length 4 -> 8
- Android: queries add launcher intent for PerAppProxyScreen
- Android: notification VISIBILITY_PRIVATE
- Android: nativeConnectServer server_id charset validation
- Android: port input range validation (1-65535)
- Android: master password Zeroizing<String> in Rust FFI
- CI: OAuth state server-side storage + one-time consume in exchange
- CI: PHP security headers (X-Content-Type-Options/X-Frame-Options)
- CI: CSP img-src remove https://*
- CI: devtools capability moved to dev.json (debug-only)
- Rust: SOCKS5 ct_eq length comparison u8 -> u32

+++ crates/core/src/proxy/socks5.rs
@@ -396,13 +396,13 @@ mod tests {
 /// Does NOT short-circuit on length mismatch — always compares the full
 ///   shorter slice to avoid leaking length information via timing.
 fn ct_eq(a: &[u8], b: &[u8]) -> bool {
-    let mut diff = 0u8;
-    // XOR length difference so different lengths are detected
-    diff ^= (a.len() as u8).wrapping_sub(b.len() as u8);
+    let mut diff = 0u32;
+    // Compare full length difference (L-7: was truncated to u8, now u32)
+    diff ^= (a.len() as u32).wrapping_sub(b.len() as u32);
     // Compare up to the shorter length
     let min_len = a.len().min(b.len());
     for i in 0..min_len {
-        diff |= a[i] ^ b[i];
+        diff |= (a[i] ^ b[i]) as u32;
     }
     diff == 0
 }
